feat: PostgreSQL database
Instead of using the SQLite database, which was used previously becaue of how it was done in the course, we are moving to a PostgreSQL database.

Created a PostgreSQL database in the VM. Created a user for this called 'rest_test'. Will have to change once deployed. All this was done for the purpose of learning how to do it.

Added a config.py file (which seems to be the default way of storing Flask config variables) and added the database uri to it.